pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "CLI AI agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
